+++ graphix/command.py
@@ -4,9 +4,10 @@ from __future__ import annotations
 
 import dataclasses
 import enum
+import math
 import sys
 from enum import Enum
-from typing import ClassVar, Literal, Union
+from typing import ClassVar, Literal, SupportsFloat, Union
 
 import numpy as np
 
@@ -18,6 +19,7 @@ from graphix.measurements import Domains
 # Ruff suggests to move this import to a type-checking block, but dataclass requires it here
 from graphix.parameter import ExpressionOrFloat  # noqa: TC001
 from graphix.pauli import Pauli
+from graphix.pretty_print import SUBSCRIPTS, SUPERSCRIPTS, OutputFormat, angle_to_str, domain_to_str
 from graphix.states import BasicStates, State
 from graphix.utils import DataclassPrettyPrintMixin
 
@@ -35,6 +37,96 @@ class CommandKind(Enum):
     Z = enum.auto()
     S = enum.auto()
     T = enum.auto()
+
+
+def command_to_str(cmd: Command, output: OutputFormat) -> str:
+    """Return the string representation of a command according to the given format.
+
+    Parameters
+    ----------
+    cmd: Command
+        The command to pretty print.
+    output: OutputFormat
+        The expected format.
+    """
+    out = [cmd.kind.name]
+
+    if cmd.kind == CommandKind.E:
+        u, v = cmd.nodes
+        if output == OutputFormat.LaTeX:
+            out.append(f"_{{{u},{v}}}")
+        elif output == OutputFormat.Unicode:
+            u_subscripts = str(u).translate(SUBSCRIPTS)
+            v_subscripts = str(v).translate(SUBSCRIPTS)
+            out.append(f"{u_subscripts}₋{v_subscripts}")
+        else:
+            out.append(f"({u},{v})")
+    elif cmd.kind == CommandKind.T:
+        pass
+    else:
+        # All other commands have a field `node` to print, together
+        # with some other arguments and/or domains.
+        arguments = []
+        if cmd.kind == CommandKind.M:
+            if cmd.plane != Plane.XY:
+                arguments.append(cmd.plane.name)
+            # We use `SupportsFloat` since `isinstance(cmd.angle, float)`
+            # is `False` if `cmd.angle` is an integer.
+            if isinstance(cmd.angle, SupportsFloat):
+                angle = float(cmd.angle)
+                if not math.isclose(angle, 0.0):
+                    arguments.append(angle_to_str(angle, output))
+            else:
+                # If the angle is a symbolic expression, we can only delegate the printing
+                # TODO: We should have a mean to specify the format
+                arguments.append(str(cmd.angle * math.pi))
+        elif cmd.kind == CommandKind.C:
+            arguments.append(str(cmd.clifford))
+        # Use of `==` here for mypy
+        command_domain = (
+            cmd.domain
+            if cmd.kind == CommandKind.X  # noqa: PLR1714
+            or cmd.kind == CommandKind.Z
+            or cmd.kind == CommandKind.S
+            else None
+        )
+        if output == OutputFormat.LaTeX:
+            out.append(f"_{{{cmd.node}}}")
+            if arguments:
+                out.append(f"^{{{','.join(arguments)}}}")
+        elif output == OutputFormat.Unicode:
+            node_subscripts = str(cmd.node).translate(SUBSCRIPTS)
+            out.append(f"{node_subscripts}")
+            if arguments:
+                out.append(f"({','.join(arguments)})")
+        else:
+            arguments = [str(cmd.node), *arguments]
+            if command_domain:
+                arguments.append(domain_to_str(command_domain))
+                command_domain = None
+            out.append(f"({','.join(arguments)})")
+        if cmd.kind == CommandKind.M and (cmd.s_domain or cmd.t_domain):
+            out = ["[", *out, "]"]
+            if cmd.t_domain:
+                if output == OutputFormat.LaTeX:
+                    t_domain_str = f"{{}}_{{{','.join(str(node) for node in cmd.t_domain)}}}"
+                elif output == OutputFormat.Unicode:
+                    t_domain_subscripts = [str(node).translate(SUBSCRIPTS) for node in cmd.t_domain]
+                    t_domain_str = "₊".join(t_domain_subscripts)
+                else:
+                    t_domain_str = f"{{{','.join(str(node) for node in cmd.t_domain)}}}"
+                out = [t_domain_str, *out]
+            command_domain = cmd.s_domain
+        if command_domain:
+            if output == OutputFormat.LaTeX:
+                domain_str = f"^{{{','.join(str(node) for node in command_domain)}}}"
+            elif output == OutputFormat.Unicode:
+                domain_superscripts = [str(node).translate(SUPERSCRIPTS) for node in command_domain]
+                domain_str = "⁺".join(domain_superscripts)
+            else:
+                domain_str = f"{{{','.join(str(node) for node in command_domain)}}}"
+            out.append(domain_str)
+    return f"{''.join(out)}"
 
 
 class _KindChecker:
